scripts/llm-evaluation/requirements.txt
ollama>=0.3.0
openai>=1.0.0
psutil>=5.9.0
python-dotenv>=1.0.0
